+++ mupen64plus-core/src/r4300/x86_64/gcop1_d.c
@@ -40,7 +40,7 @@ void genadd_d(void)
    inc_m32rel(&instr_count[119]);
 #endif
 #ifdef INTERPRET_ADD_D
-    gencallinterp((unsigned long long)cached_interpreter_table.ADD_D, 0);
+    gencallinterp((uint64_t)cached_interpreter_table.ADD_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -58,7 +58,7 @@ void gensub_d(void)
    inc_m32rel(&instr_count[120]);
 #endif
 #ifdef INTERPRET_SUB_D
-   gencallinterp((unsigned long long)cached_interpreter_table.SUB_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.SUB_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -76,7 +76,7 @@ void genmul_d(void)
    inc_m32rel(&instr_count[121]);
 #endif
 #ifdef INTERPRET_MUL_D
-   gencallinterp((unsigned long long)cached_interpreter_table.MUL_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.MUL_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -94,7 +94,7 @@ void gendiv_d(void)
    inc_m32rel(&instr_count[122]);
 #endif
 #ifdef INTERPRET_DIV_D
-   gencallinterp((unsigned long long)cached_interpreter_table.DIV_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.DIV_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -112,7 +112,7 @@ void gensqrt_d(void)
    inc_m32rel(&instr_count[123]);
 #endif
 #ifdef INTERPRET_SQRT_D
-   gencallinterp((unsigned long long)cached_interpreter_table.SQRT_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.SQRT_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -129,7 +129,7 @@ void genabs_d(void)
    inc_m32rel(&instr_count[124]);
 #endif
 #ifdef INTERPRET_ABS_D
-   gencallinterp((unsigned long long)cached_interpreter_table.ABS_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.ABS_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -146,7 +146,7 @@ void genmov_d(void)
    inc_m32rel(&instr_count[125]);
 #endif
 #ifdef INTERPRET_MOV_D
-   gencallinterp((unsigned long long)cached_interpreter_table.MOV_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.MOV_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -164,7 +164,7 @@ void genneg_d(void)
    inc_m32rel(&instr_count[126]);
 #endif
 #ifdef INTERPRET_NEG_D
-   gencallinterp((unsigned long long)cached_interpreter_table.NEG_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.NEG_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -181,7 +181,7 @@ void genround_l_d(void)
    inc_m32rel(&instr_count[127]);
 #endif
 #ifdef INTERPRET_ROUND_L_D
-   gencallinterp((unsigned long long)cached_interpreter_table.ROUND_L_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.ROUND_L_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&round_mode);
@@ -199,7 +199,7 @@ void gentrunc_l_d(void)
    inc_m32rel(&instr_count[128]);
 #endif
 #ifdef INTERPRET_TRUNC_L_D
-   gencallinterp((unsigned long long)cached_interpreter_table.TRUNC_L_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TRUNC_L_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&trunc_mode);
@@ -217,7 +217,7 @@ void genceil_l_d(void)
    inc_m32rel(&instr_count[129]);
 #endif
 #ifdef INTERPRET_CEIL_L_D
-   gencallinterp((unsigned long long)cached_interpreter_table.CEIL_L_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CEIL_L_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&ceil_mode);
@@ -235,7 +235,7 @@ void genfloor_l_d(void)
    inc_m32rel(&instr_count[130]);
 #endif
 #ifdef INTERPRET_FLOOR_L_D
-   gencallinterp((unsigned long long)cached_interpreter_table.FLOOR_L_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.FLOOR_L_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&floor_mode);
@@ -253,7 +253,7 @@ void genround_w_d(void)
    inc_m32rel(&instr_count[127]);
 #endif
 #ifdef INTERPRET_ROUND_W_D
-   gencallinterp((unsigned long long)cached_interpreter_table.ROUND_W_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.ROUND_W_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&round_mode);
@@ -271,7 +271,7 @@ void gentrunc_w_d(void)
    inc_m32rel(&instr_count[128]);
 #endif
 #ifdef INTERPRET_TRUNC_W_D
-   gencallinterp((unsigned long long)cached_interpreter_table.TRUNC_W_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TRUNC_W_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&trunc_mode);
@@ -289,7 +289,7 @@ void genceil_w_d(void)
    inc_m32rel(&instr_count[129]);
 #endif
 #ifdef INTERPRET_CEIL_W_D
-   gencallinterp((unsigned long long)cached_interpreter_table.CEIL_W_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CEIL_W_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&ceil_mode);
@@ -307,7 +307,7 @@ void genfloor_w_d(void)
    inc_m32rel(&instr_count[130]);
 #endif
 #ifdef INTERPRET_FLOOR_W_D
-   gencallinterp((unsigned long long)cached_interpreter_table.FLOOR_W_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.FLOOR_W_D, 0);
 #else
    gencheck_cop1_unusable();
    fldcw_m16rel((unsigned short*)&floor_mode);
@@ -325,7 +325,7 @@ void gencvt_s_d(void)
    inc_m32rel(&instr_count[117]);
 #endif
 #ifdef INTERPRET_CVT_S_D
-   gencallinterp((unsigned long long)cached_interpreter_table.CVT_S_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CVT_S_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -341,7 +341,7 @@ void gencvt_w_d(void)
    inc_m32rel(&instr_count[117]);
 #endif
 #ifdef INTERPRET_CVT_W_D
-   gencallinterp((unsigned long long)cached_interpreter_table.CVT_W_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CVT_W_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -357,7 +357,7 @@ void gencvt_l_d(void)
    inc_m32rel(&instr_count[117]);
 #endif
 #ifdef INTERPRET_CVT_L_D
-   gencallinterp((unsigned long long)cached_interpreter_table.CVT_L_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CVT_L_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.fs]));
@@ -373,7 +373,7 @@ void genc_f_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_F_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_F_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_F_D, 0);
 #else
    gencheck_cop1_unusable();
    and_m32rel_imm32((unsigned int*)&FCR31, ~0x800000);
@@ -386,7 +386,7 @@ void genc_un_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_UN_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_UN_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_UN_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -408,7 +408,7 @@ void genc_eq_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_EQ_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_EQ_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_EQ_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -430,7 +430,7 @@ void genc_ueq_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_UEQ_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_UEQ_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_UEQ_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -453,7 +453,7 @@ void genc_olt_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_OLT_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_OLT_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_OLT_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -475,7 +475,7 @@ void genc_ult_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_ULT_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_ULT_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_ULT_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -498,7 +498,7 @@ void genc_ole_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_OLE_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_OLE_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_OLE_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -520,7 +520,7 @@ void genc_ule_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_ULE_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_ULE_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_ULE_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -543,7 +543,7 @@ void genc_sf_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_SF_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_SF_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_SF_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -562,7 +562,7 @@ void genc_ngle_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_NGLE_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_NGLE_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_NGLE_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -584,7 +584,7 @@ void genc_seq_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_SEQ_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_SEQ_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_SEQ_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -606,7 +606,7 @@ void genc_ngl_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_NGL_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_NGL_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_NGL_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -629,7 +629,7 @@ void genc_lt_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_LT_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_LT_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_LT_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -651,7 +651,7 @@ void genc_nge_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_NGE_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_NGE_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_NGE_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -674,7 +674,7 @@ void genc_le_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_LE_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_LE_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_LE_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
@@ -696,7 +696,7 @@ void genc_ngt_d(void)
    inc_m32rel(&instr_count[118]);
 #endif
 #ifdef INTERPRET_C_NGT_D
-   gencallinterp((unsigned long long)cached_interpreter_table.C_NGT_D, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.C_NGT_D, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_double[dst->f.cf.ft]));
