+++ mupen64plus-core/src/r4300/x86_64/gtlb.c
@@ -38,7 +38,7 @@ void gentlbwi(void)
 #if defined(COUNT_INSTR)
    inc_m32rel(&instr_count[104]);
 #endif
-   gencallinterp((unsigned long long)cached_interpreter_table.TLBWI, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TLBWI, 0);
    /*dst->local_addr = code_length;
    mov_m32_imm32((void *)(&PC), (unsigned int)(dst));
    mov_reg32_imm32(EAX, (unsigned int)(TLBWI));
@@ -51,7 +51,7 @@ void gentlbp(void)
 #if defined(COUNT_INSTR)
    inc_m32rel(&instr_count[105]);
 #endif
-   gencallinterp((unsigned long long)cached_interpreter_table.TLBP, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TLBP, 0);
    /*dst->local_addr = code_length;
    mov_m32_imm32((void *)(&PC), (unsigned int)(dst));
    mov_reg32_imm32(EAX, (unsigned int)(TLBP));
@@ -64,7 +64,7 @@ void gentlbr(void)
 #if defined(COUNT_INSTR)
    inc_m32rel(&instr_count[106]);
 #endif
-   gencallinterp((unsigned long long)cached_interpreter_table.TLBR, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TLBR, 0);
    /*dst->local_addr = code_length;
    mov_m32_imm32((void *)(&PC), (unsigned int)(dst));
    mov_reg32_imm32(EAX, (unsigned int)(TLBR));
@@ -77,7 +77,7 @@ void generet(void)
 #if defined(COUNT_INSTR)
    inc_m32rel(&instr_count[108]);
 #endif
-   gencallinterp((unsigned long long)cached_interpreter_table.ERET, 1);
+   gencallinterp((uint64_t)cached_interpreter_table.ERET, 1);
    /*dst->local_addr = code_length;
    mov_m32_imm32((void *)(&PC), (unsigned int)(dst));
    genupdate_system(0);
@@ -92,6 +92,6 @@ void gentlbwr(void)
 #if defined(COUNT_INSTR)
    inc_m32rel(&instr_count[107]);
 #endif
-   gencallinterp((unsigned long long)cached_interpreter_table.TLBWR, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.TLBWR, 0);
 }
